backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
openai>=1.0.0
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
google-genai>=1.0.0
numpy>=1.26.0
sse-starlette>=2.0.0
ultralytics>=8.3.0
opencv-python-headless>=4.9.0
supabase>=2.0.0
